+++ Sources/UINotificationStyles/UINotificationSystemStyle.swift
@@ -10,9 +10,11 @@ import UIKit
 
 /// Defines the default style for notifications.
 public struct UINotificationSystemStyle: UINotificationStyle {
-    public var font: UIFont = UIFont.systemFont(ofSize: 13, weight: .semibold)
+    public var titleFont: UIFont = UIFont.systemFont(ofSize: 13, weight: .semibold)
+    public var subtitleFont: UIFont = UIFont.systemFont(ofSize: 13, weight: .regular)
     public var backgroundColor: UIColor = UIColor.white
-    public var textColor: UIColor = UIColor.black
+    public var titleTextColor: UIColor = UIColor.black
+    public var subtitleTextColor: UIColor = UIColor.darkGray
     public var height: UINotificationHeight = .navigationBar
     public var interactive: Bool = true
     public var chevronImage: UIImage?